bll/manager/requirements.txt
Flask==2.0.2
Flask_Cors==3.0.10
PyJWT==2.8.0
python_bcrypt==0.3.2
SQLAlchemy==2.0.30
